pyproject.toml
[tool.poetry]
name = "drf-text-store"
version = "0.1.0"
description = "Django REST Framework Text Store"
license = "MIT"
authors = ["Matt McGrattan <matt.mcgrattan@digirati.com", "Finlay McCourt <finlay.mccourt@digirati.com>"]
readme = "README.md"
packages = [
    { include = "text_store" },
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "4.0.2"
djangorestframework = "^3.13.1"
django-model-utils = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
requests = "^2.27.1"
pytest-docker = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
